fix(llm-proxy): review findings — migration donor filters, managed-row guards, copy sweep
