+++ platform/frontend/src/app/llm/proxy/oauth-clients/page.tsx
@@ -9,7 +9,7 @@ import { Copy, Pencil, Plus, RefreshCw, Trash2 } from "lucide-react";
 import { useCallback, useEffect, useState } from "react";
 import { toast } from "sonner";
 import { ErrorBoundary } from "@/app/_parts/error-boundary";
-import { useSetLlmProxyAction } from "@/app/llm/proxy/layout";
+import { useSetLlmProxyAction } from "@/app/llm/proxy/_parts/llm-proxy-action-context";
 import { CreateOAuthClientDialog } from "@/components/create-oauth-client-dialog";
 import { DeleteConfirmDialog } from "@/components/delete-confirm-dialog";
 import {
